Spike method for dynamically generating propTypes
`propTypes.ts` implements a method for using react-docgen-typescript to map out a component's properties and tranfsorming them to React.PropTypes. This works if the propTypes are attached to the component in the index.ts file. The Badge is used as an example.

The downside of this method is that the propTypes have to be read from the file on disk and processed once every time the component is imported. This takes about 1200 ms on a fast MBP. An ideal solution would either involve altering the source at compilation via some webpack loader.